services/data-ingestion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
redis==5.0.1
python-dotenv==1.0.0
aiohttp==3.13.3
numpy==1.26.3
